Fix 5 remaining benchmark API issues
- bench_scaled_fp4_experts_quant.py: input_global_scale must be 1D tensor (reshape to [1])
- bench_cutlass_fp4_group_mm.py: expert_offsets must have same size as problem_sizes
  (changed from [num_experts+1] to [num_experts])
- bench_apply_shuffle_mul_sum.py: factors dtype must match output (convert to bfloat16)
- bench_trtllm_fp4_block_scale_moe.py: increase MIN_DECODE_BATCH to 8 (B=2 still fails)
- bench_fused_moe_kernel.py: fix import path to fused_moe_triton.fused_moe.fused_moe

+++ deepseek_kernel_benchmarks/scripts/bench_apply_shuffle_mul_sum.py
@@ -53,8 +53,9 @@ def bench_apply_shuffle_mul_sum(sgl_kernel, B: int, S: int, hidden_size: int,
     permutation = torch.arange(tokens, dtype=torch.int32, device=device).repeat_interleave(topk)
 
     # Factors: weights for each expert output (softmax of routing scores)
+    # Must match output dtype (bfloat16)
     factors = torch.randn(total_expert_tokens, dtype=torch.float32, device=device)
-    factors = factors.view(tokens, topk).softmax(dim=-1).flatten()
+    factors = factors.view(tokens, topk).softmax(dim=-1).flatten().to(torch.bfloat16)
 
     def kernel_fn():
         apply_shuffle_mul_sum(input_tensor, output, permutation, factors)
